+++ test/corelib/src/utility/to_number_tests.cpp
@@ -2,11 +2,13 @@
 // Distributed under Boost license
 
 #include <jsoncons/utility/to_number.hpp>
+#include <clocale>
+#include <iostream>
 #include <catch/catch.hpp>
 
 using namespace jsoncons;
 
-TEST_CASE("utility::to_integer tests")
+TEST_CASE("jsoncons::utility::to_integer tests")
 {
     SECTION("")
     {
@@ -162,3 +164,54 @@ TEST_CASE("utility::to_integer_unchecked tests")
     }
 }
 
+TEST_CASE("jsoncons::utility::to_double tests")
+{
+    SECTION("minus overflow")
+    {
+        std::string s = "-50011731000E95978";
+        double val;
+        auto result = jsoncons::utility::to_double(s.data(), s.length(), val);
+        REQUIRE(result.ec == std::errc::result_out_of_range);
+        auto result2 = jsoncons::utility::to_double(s.c_str(), s.length(), val);
+        REQUIRE(result2.ec == std::errc::result_out_of_range);
+    }
+    SECTION("overflow")
+    {
+        std::wstring s = L"50011731000E95978";
+        double val;
+        auto result = jsoncons::utility::to_double(s.data(), s.length(), val);
+        REQUIRE(result.ec == std::errc::result_out_of_range);
+        auto result2 = jsoncons::utility::to_double(s.c_str(), s.length(), val);
+        REQUIRE(result2.ec == std::errc::result_out_of_range);
+    }
+    SECTION("locale")
+    {
+        std::string prev_loc = std::setlocale(LC_ALL, nullptr);
+        if (std::setlocale(LC_NUMERIC, "de_DE.UTF-8"))
+        {
+            std::string s = "1234.5678";
+            std::wstring ws = L"1234.5678";
+            
+            double val1{0};
+            auto result1 = jsoncons::utility::to_double(&s[0], s.length(), val1);
+            CHECK(result1);
+            CHECK(1234.5678 == val1);
+
+            double val2{ 0 };
+            auto result2 = jsoncons::utility::to_double(&ws[0], ws.length(), val2);
+            CHECK(result2);
+            CHECK(1234.5678 == val2);
+            
+            double val3{0};
+            auto result3 = jsoncons::utility::to_double(s.c_str(), s.length(), val3);
+            CHECK(result3);
+            CHECK(1234.5678 == val3);
+
+            double val4{ 0 };
+            auto result4 = jsoncons::utility::to_double(ws.c_str(), ws.length(), val4);
+            CHECK(result4);
+            CHECK(1234.5678 == val4);
+        }
+        std::setlocale(LC_ALL, prev_loc.c_str());
+    }
+}